Correctly escape variables when templating GraphiQL response
This commit makes sure that values are escaped correctly when doing
re.sub() while rendering a GraphiQL response. This is necessary when
non-ASCII characters end up in the GraphQL variables, for example an
umlaut. json.dumps() encodes such characters using a unicode escape
sequence starting with "\u", which isn't a valid regex. Hence, we would
get this error:

    re.error: bad escape \u at position 11

We get around this by re.escape()-ing everything before doing anything
regexy with it.